+++ CHANGES.rst
@@ -12,6 +12,9 @@ Changelog
   a requirements text file now as well
 - added `pww_version()` function to `weka.core.version` to obtain the installed
   version of python-weka-wrapper3
+- added sub-command `bootstrap` to the `pww-packages` command-line tool to output a
+  Python script that can reinstall python-weka-wrapper3 and all the packages in
+  another virtual environment or even another machine
 
 
 0.3.1 (2024-07-17)